# docs/swagger.yaml
basePath: /
definitions:
...
  emplacc-api_internal_dto_request.ProblemCreateRequest:
    properties:
      creator_id:
        type: string
      description:
        items:
          type: string
        type: array
      name:
        type: string
    type: object
  emplacc-api_internal_dto_request.ProblemUpdateRequest:
    properties:
      description:
        items:
          type: string
        type: array
      name:
        type: string
    type: object
...
  emplacc-api_internal_dto_response.ProblemListResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      problems:
        items:
          $ref: '#/definitions/emplacc-api_internal_dto_response.ProblemResponse'
        type: array
      total_count:
        type: integer
    type: object
  emplacc-api_internal_dto_response.ProblemResponse:
    properties:
      created_at:
        type: string
      creator_id:
        type: string
      description:
        items:
          type: string
        type: array
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  emplacc-api_internal_dto_response.ProblemUniversalResponse:
    properties:
      id:
        type: string
      message:
        type: string
    type: object
  emplacc-api_internal_dto_response.ProblemsByUserId:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      problems:
        items:
          $ref: '#/definitions/emplacc-api_internal_dto_response.ProblemResponse'
        type: array
      total_count:
        type: integer
      user_id:
        type: string
    type: object
...
info:
  contact: {}
  description: API для Emplacc.
  title: Emplacc API
  version: "1.0"
paths:
...
  /problem:
    post:
      consumes:
      - application/json
      description: Создает новую проблему с указанными параметрами
      parameters:
      - description: Данные для создания проблемы
        in: body
        name: problem
        required: true
        schema:
          $ref: '#/definitions/emplacc-api_internal_dto_request.ProblemCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Проблема успешно создана
          schema:
            $ref: '#/definitions/emplacc-api_internal_dto_response.ProblemUniversalResponse'
        "400":
          description: Ошибка в запросе или некорректный идентификатор пользователя
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Нет или неверный токен
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера при создании проблемы
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создание новой проблемы
      tags:
      - Problems
  /problem/{id}:
    delete:
      consumes:
      - application/json
      description: Логическое удаление проблемы по ID, включая связанные данные (поле
        deleted = true)
      parameters:
      - description: ID проблемы
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Проблема успешно удалена
          schema:
            $ref: '#/definitions/emplacc-api_internal_dto_response.ProblemUniversalResponse'
        "404":
          description: Проблема не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера при удалении проблемы
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление проблемы
      tags:
      - Problems
    get:
      consumes:
      - application/json
      description: Получает данные проблемы по её уникальному идентификатору
      parameters:
      - description: ID проблемы
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Проблема успешно получена
          schema:
            $ref: '#/definitions/emplacc-api_internal_dto_response.ProblemResponse'
        "400":
          description: Некорректный идентификатор проблемы
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Нет или неверный токен
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Проблема не найдена
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера при получении проблемы
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение проблемы по ID
      tags:
      - Problems
    patch:
      consumes:
      - application/json
      description: Обновляет данные проблемы по её ID
      parameters:
      - description: ID проблемы
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления проблемы
        in: body
        name: problem
        required: true
        schema:
          $ref: '#/definitions/emplacc-api_internal_dto_request.ProblemUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Проблема успешно обновлена
          schema:
            $ref: '#/definitions/emplacc-api_internal_dto_response.ProblemUniversalResponse'
        "400":
          description: Некорректный идентификатор или ошибка в запросе
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Нет или неверный токен
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера при обновлении проблемы
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновление проблемы
      tags:
      - Problems
  /problem/all/{page}/{pagesize}:
    get:
      consumes:
      - application/json
      description: Получает список всех проблем с учетом пагинации, исключая удаленные
      parameters:
      - description: Номер страницы
        in: path
        name: page
        required: true
        type: integer
      - description: Размер страницы
        in: path
        name: pagesize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список проблем успешно получен
          schema:
            $ref: '#/definitions/emplacc-api_internal_dto_response.ProblemListResponse'
        "400":
          description: Ошибка в запросе
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Нет или неверный токен
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера при получении проблем
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение списка всех проблем
      tags:
      - Problems
  /problem/user/{id}/{page}/{pagesize}:
    get:
      consumes:
      - application/json
      description: Получает список проблем, созданных указанным пользователем, с учетом
        пагинации
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: string
      - description: Номер страницы
        in: path
        name: page
        required: true
        type: integer
      - description: Размер страницы
        in: path
        name: pagesize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список проблем успешно получен
          schema:
            $ref: '#/definitions/emplacc-api_internal_dto_response.ProblemsByUserId'
        "400":
          description: Ошибка в запросе
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Нет или неверный токен
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера при получении проблем
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение проблем по ID пользователя
      tags:
      - Problems
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"